fixed amount of training per epoch, removed spaces from filenames
